rename tokensBuffer to dentsBuffer
The buffer only contains DEDENTs (and, for one edge case, INDENTs).
We can't use it for everything, because the non-NL tokens (from the
fast forwarding feature, to ignore blank lines) need to come _after_
any inserted/buffered dents.

The previous commit also moved the EOF to `nextNonNL`, for consistency
(and to allow this rename) more than anything else.