add examples project, change dedent behavior
Dedents are now preceded by a newline token. There is only one nl per
string of dedents. For instance, if you had:

    foo
      bar
         baz
    hello

Then there are two dedents between `baz` and `hello`. This would come
through as `newline, dedent, dedent`.

This change in behavior is to enable behavior as shown in the example,
which is just a simple indent-based calculator language. I anticipate
that this use case will be the norm, rather than a contrived example,
which is why I'm letting the test dictate behavior.