add a helper class to "fork" grammars
It takes a template grammar and spits out two `.g4` files, one for
a denting version, and one for a "braced" version. The braces used are
unicode triangles to avoid conflict with typical languages. The idea
is to create two version of the same grammar, one with denting and one
without, so that we can benchmark them. This approach is a bit hacky,
but it's simple and should be enough.