have the rewriter explicitly indent
it only uses the incoming token's position within each line, and only
for non-dent/nl tokens. For dedents, it explicitly adds indentation
before the braces by keeping track of the nesting count and assuming
two spaces per indent. This is still not perfect, but if the input
source does have two spaces per indent level, we actually get pretty
much exactly what we want. This is now a fair comparison.