fix benchmark grammar, I think
Discovered while looking into bug #22. Before, I had a block as an
indent, and then statements separated by newlines. Instead, I think it
needs to just be a series of statements, with any one-liner statements
(of which the only is `assign` having their own newline).

This is because antlr-denter will insert a single newline before a
stretch of dedents.